requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
gunicorn==21.2.0
onnxruntime==1.16.3
numpy==1.24.3
pillow==10.0.1
python-multipart==0.0.6
opencv-python-headless==4.8.1.78